requirements.txt
click==7.0 \
    --hash=sha256:2335065e6395b9e67ca716de5f7526736bfa6ceead690adf616d925bdc622b13 \
    --hash=sha256:5b94b49521f6456670fdb30cd82a4eca9412788a93fa6dd6df72c94d5a8ff2d7
pymql5==1.2.0 \
    --hash=sha256:ef5a93d44b8bdb68327ee45e3dc4216788f287f926ab7c765cb2689f526d6155 \
    --hash=sha256:3a550ed761664850d42131649a2c0fead0457b8245c5045e72cfca690208ed7c
python-dotenv==0.10.3 \
    --hash=sha256:f157d71d5fec9d4bd5f51c82746b6344dffa680ee85217c123f4a0c8117c4544 \
    --hash=sha256:debd928b49dbc2bf68040566f55cdb3252458036464806f4094487244e2a4093
